+++ BlazorElectronics/Shared/Mutual/CategoryIdMap.cs
@@ -2,12 +2,12 @@ namespace BlazorElectronics.Shared.Mutual;
 
 public sealed class CategoryIdMap
 {
-    public CategoryIdMap( int tier, int categoryId )
+    public CategoryIdMap( int tier, short categoryId )
     {
         Tier = tier;
         CategoryId = categoryId;
     }
     
     public int Tier { get; set; }
-    public int CategoryId { get; set; }
+    public short CategoryId { get; set; }
 }
