+++ BlazorElectronics/Shared/Outbound/Categories/CategoriesResponse.cs
@@ -2,7 +2,7 @@ namespace BlazorElectronics.Shared.Outbound.Categories;
 
 public sealed class CategoriesResponse
 {
-    public Dictionary<int, PrimaryCategoryResponse> PrimaryCategories { get; set; } = new();
-    public Dictionary<int, SecondaryCategoryResponse> SecondaryCategories { get; set; } = new();
-    public Dictionary<int, TertiaryCategoryResponse> TertiaryCategories { get; set; } = new();
+    public Dictionary<short, PrimaryCategoryResponse> PrimaryCategories { get; set; } = new();
+    public Dictionary<short, SecondaryCategoryResponse> SecondaryCategories { get; set; } = new();
+    public Dictionary<short, TertiaryCategoryResponse> TertiaryCategories { get; set; } = new();
 }
